fix: address PR review feedback — split vrm_ops.rs and use RequestDetachVrm
FB1: Split monolithic vrm_ops.rs (16 handlers, 414 lines) into vrm/ directory
with 7 operation-specific submodules: expressions, vrma, position, bone, look,
spring_bones, speech. Rename vrm.rs to vrm/mod.rs (attach/detach stays).

FB2: Update bevy_vrm1 from 0.6.3 to 0.6.4 and replace manual component
removal in vrm_detach.rs with RequestDetachVrm trigger, which properly
cleans up all VRM components, materials, and bone entities.
